LayoutTests/fast/events/wheel/wheel-events-become-non-cancelable-expected.txt

Tests that events in the gesture are non-cancelable if preventDefault was not called on the first event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstWasCancelable is true
PASS becameNonCancelable is true
PASS successfullyParsed is true

TEST COMPLETE
